refactor: extract useCloudState hook from AppContext
Move 15 elizaCloud* useState calls + 8 refs + 3 callbacks (pollCloudCredits,
handleCloudLogin, handleCloudDisconnect) + cloud auth effect into
useCloudState. Returns pollCloudCredits for startup/WS handler wiring.

AppContext.tsx: 7,439 → 6,992 lines (-447)
useCloudState.ts: 645 lines (new)
Cumulative: 8,366 → 6,992 (-1,374 lines, 7 hooks extracted)
